Fix warning C4244 on MSVC about conversion from double to float
This is a minor change to avoid getting several warnings about
converting from double to float when we use MSVC compiler.

+++ include/private/SkFloatingPoint.h
@@ -240,7 +240,7 @@ static inline double sk_ieee_double_divide(double numer, double denom) {
 static inline float sk_ieee_float_divide_TODO_IS_DIVIDE_BY_ZERO_SAFE_HERE(float n, float d) {
     return sk_ieee_float_divide(n,d);
 }
-static inline float sk_ieee_double_divide_TODO_IS_DIVIDE_BY_ZERO_SAFE_HERE(double n, double d) {
+static inline double sk_ieee_double_divide_TODO_IS_DIVIDE_BY_ZERO_SAFE_HERE(double n, double d) {
     return sk_ieee_double_divide(n,d);
 }
